Finishing up consolidating code
