+++ src/applications/static-pages/find-forms/tests/actions/index.unit.spec.js
@@ -4,7 +4,6 @@ import {
   fetchFormsAction,
   fetchFormsFailure,
   fetchFormsSuccess,
-  fetchFormsThunk,
   updatePaginationAction,
   updateResults,
   updateSortByPropertyName,
@@ -116,70 +115,6 @@ describe('Find VA Forms actions', () => {
           type: UPDATE_RESULTS,
         }),
       ).to.be.true;
-    });
-  });
-
-  describe('fetchFormsThunk', () => {
-    let mockedLocation;
-    let mockedHistory;
-
-    beforeEach(() => {
-      mockedLocation = {
-        search: '',
-        pathname: '',
-      };
-
-      mockedHistory = {
-        replaceState: sinon.stub(),
-      };
-    });
-
-    it('updates search params', async () => {
-      const dispatch = sinon.stub();
-      const query = 'health';
-      const thunk = fetchFormsThunk(query, {
-        location: mockedLocation,
-        history: mockedHistory,
-        mockRequest: true,
-      });
-
-      await thunk(dispatch);
-
-      const replaceStateStub = mockedHistory.replaceState;
-
-      expect(replaceStateStub.calledOnce).to.be.true;
-      expect(replaceStateStub.firstCall.args[2]).to.be.equal('?q=health');
-    });
-
-    it('calls dispatch', async () => {
-      const dispatch = sinon.stub();
-      const query = 'health';
-      const thunk = fetchFormsThunk(query, {
-        location: mockedLocation,
-        history: mockedHistory,
-        mockRequest: true,
-      });
-
-      await thunk(dispatch);
-
-      expect(
-        dispatch.firstCall.calledWith({
-          type: FETCH_FORMS,
-          query,
-        }),
-      ).to.be.true;
-
-      expect(
-        dispatch.secondCall.calledWith({
-          page: 1,
-          startIndex: 0,
-          type: UPDATE_PAGINATION,
-        }),
-      ).to.be.true;
-
-      const thirdCallAction = dispatch.thirdCall.args[0];
-      expect(thirdCallAction.type).to.be.equal(FETCH_FORMS_SUCCESS);
-      expect(thirdCallAction.results).to.be.an('array');
     });
   });
 });
